[FIX/VG-446] hlsl에 불필요한 casting 제거

+++ Source/GA6thFinal_Framework/Shaders/vs_particle_quad.hlsl
@@ -32,7 +32,7 @@ VSOutput vs_main(VSInput input)
     
     output.color = instanceInfo.Color;
     output.emitterIndex.x = instanceInfo.EmitterIndex;
-    output.emitterIndex.y = (int) instanceInfo.Paddings2.x;
+    output.emitterIndex.y = instanceInfo.Paddings2.x;
 
     if(instanceInfo.FrameInfo.x < 0)
     {
